refactor(frontend): pages/consumer/ 直下の平置き hook 8件を model/ へ移設する（#1047）
fleet 裁定(a)（2026-08-04）＝平置き hook は「hook は model/」条項の違反。
2条項は衝突していない: 「サブディレクトリは5語のみ」は名前の禁止(negative)、
「hook は model/」は置き場の指定(positive)で、前者の射程外でも後者は単独で効く。

- 5実装＋併置 test 3件＝8ファイルを model/ へ。内訳 = **R100 が 7件**（中身バイト不変）／
  **R097 が 1件**（use-spa-link-interception.ts のみ・下記の出ていく import 補正のため）
  use-consumer-motion / use-consumer-theme / use-public-document-title /
  use-spa-link-interception / use-theme-preview-bridge
- import 追随（入ってくる側）: スライス直下4ファイル
  （PublicSiteShell / PublicRecordDetailPage / PublicBrowsePage / PublicShell）
- import 補正（**出ていく側**）: 1件 — use-spa-link-interception.ts の
  ./resolve-spa-link → ../resolve-spa-link。1階層降りたことで壊れた分。
  移設8ファイルの ./ 参照を機械監査し、該当は1件のみと確認（残り3件は一緒に移動した
  相棒を指す test で同階層のままが正）

裁定条件②（use-* を export するモジュールのみ）は事前に実測済みで、5実装すべてが
use* 関数を export＝8/8 が射程内だった。外部参照は 0（すべて pages/consumer 内で閉じている）。

これで pages/consumer に平置き hook は 0 件。

検証: npm run check --prefix frontend = exit 0（vitest 119 files 666 tests 含む）。

Closes #1047
