+++ frontend/src/pages/consumer/model/use-public-document-title.ts
