+++ frontend/src/pages/consumer/model/use-theme-preview-bridge.ts
